+++ src/main/java/org/jabref/gui/fieldeditors/EditorTextArea.java
@@ -1,18 +1,25 @@
 package org.jabref.gui.fieldeditors;
 
-import java.net.URL;
-import java.util.List;
-import java.util.ResourceBundle;
-
+import com.sun.javafx.scene.control.skin.TextAreaSkin;
 import javafx.fxml.Initializable;
 import javafx.scene.control.ContextMenu;
 import javafx.scene.control.MenuItem;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
 
-import com.sun.javafx.scene.control.skin.TextAreaSkin;
+import javax.annotation.Nonnull;
+import java.net.URL;
+import java.util.List;
+import java.util.ResourceBundle;
 
 public class EditorTextArea extends javafx.scene.control.TextArea implements Initializable {
+
+    /**
+     *  Variable that contains user-defined behavior for paste action.
+     *  Set empty paste behavior by default
+     */
+    @Nonnull
+    private PasteActionHandler pasteActionHandler = new EmptyPasteHandler();
 
     public EditorTextArea() {
         this("");
@@ -64,6 +71,41 @@ public class EditorTextArea extends javafx.scene.control.TextArea implements Ini
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
+    }
 
+    /**
+     * Set pasteActionHandler variable to passed handler
+     * @param  handler an instance of PasteActionHandler that describes paste behavior
+     */
+    public void setPasteActionHandler(@Nonnull PasteActionHandler handler) {
+        this.pasteActionHandler = handler;
+    }
+
+    /**
+     *  Override javafx TextArea method applying TextArea.paste() and pasteActionHandler after
+     */
+    @Override
+    public void paste() {
+        super.paste();
+        pasteActionHandler.handle();
+    }
+
+    /**
+     *  Interface presents user-described paste behaviour applying to paste method
+     */
+
+    @FunctionalInterface
+    public interface PasteActionHandler {
+        void handle();
+    }
+
+    /**
+     * Empty interface implementation to do nothing external on paste method
+     */
+    private static class EmptyPasteHandler implements PasteActionHandler {
+        @Override
+        public void handle() {
+
+        }
     }
 }