+++ CHANGELOG.md
@@ -24,6 +24,7 @@ We refer to [GitHub issues](https://github.com/JabRef/jabref/issues) by using `#
 - We changed the default paths for the OpenOffice/LibreOffice binaries to the default path for LibreOffice
 - We no longer create a new entry editor when selecting a new entry to increase performance. [#3187](https://github.com/JabRef/jabref/pull/3187)
 - We increased performance and decreased the memory footprint of the entry editor drastically. [#3331](https://github.com/JabRef/jabref/pull/3331)
+- We added auto url formatting when user paste link to URL field in Entry editor  [#254](https://github.com/koppor/jabref/issues/254)
 
 
 ### Fixed